refactor: centralize components in design system
