+++ app/(tabs)/explore.tsx
@@ -1,12 +1,14 @@
 import { Image } from 'expo-image';
 import { Platform, StyleSheet } from 'react-native';
 
-import { Collapsible } from '@/components/Collapsible';
-import { ExternalLink } from '@/components/ExternalLink';
-import ParallaxScrollView from '@/components/ParallaxScrollView';
-import { ThemedText } from '@/components/ThemedText';
-import { ThemedView } from '@/components/ThemedView';
-import { IconSymbol } from '@/components/ui/IconSymbol';
+import {
+  Collapsible,
+  ExternalLink,
+  ParallaxScrollView,
+  ThemedText,
+  ThemedView,
+  IconSymbol,
+} from '@/design-system';
 
 export default function TabTwoScreen() {
   return (
