+++ design-system/index.ts
@@ -1,0 +1,15 @@
+export * from '../components/Collapsible';
+export * from '../components/ColorPaletteDisplay';
+export * from '../components/ErrorBoundary';
+export * from '../components/ExternalLink';
+export * from '../components/HelloWave';
+export * from '../components/HapticTab';
+export * from '../components/InspirationCard';
+export * from '../components/LoadingOverlay';
+export { default as ParallaxScrollView } from '../components/ParallaxScrollView';
+export * from '../components/PhotoUploader';
+export * from '../components/StylingTips';
+export * from '../components/ThemedText';
+export * from '../components/ThemedView';
+export * from '../components/ui/IconSymbol';
+export { default as TabBarBackground, useBottomTabOverflow } from '../components/ui/TabBarBackground';
